requirements.txt
pyttsx3
pyaudio
speechrecognition
google-generativeai
dotenv
asyncio
